requirements.txt
fastapi
uvicorn[standard]
slack_sdk
python-dotenv
openai
playwright
httpx 